Fix select CD
